job runner stopted
changed Notification block related to job runnr

+++ src/main/java/com/maan/eway/notification/bean/MailDataDetails.java
@@ -82,7 +82,7 @@ private static final long serialVersionUID = 1L;
 
     //--- ENTITY LINKS ( RELATIONSHIP )
     @Column(name="NOTIF_NO", nullable=false)
-    private Integer    notifNo ;
+    private Long    notifNo ;
     
     @Column(name="PUSHED_BY", length=100)
     private String     pushedBy ;
